move hetero graph to data.py + fixes
