+++ mlx_graphs/data/__init__.py
@@ -1,4 +1,3 @@
-from .data import GraphData  # noqa
+from .data import GraphData, HeteroGraphData  # noqa
 from .batch import GraphDataBatch, batch, unbatch  # noqa
 from .collate import collate  # noqa
-from .hetero_data import HeteroGraphData  # noqa
